refactor: move pharmacy related schemas into file
